Use atomic functions where appropriate
Also, rename atomic functions to be consistent with the rest of the
platform/threading functions, and move atomic functions to threading*
files rather than platform* files

+++ libobs/graphics/graphics-internal.h
@@ -241,5 +241,5 @@ struct graphics_subsystem {
 	DARRAY(struct vec2)    texverts[16];
 
 	pthread_mutex_t        mutex;
-	volatile int           ref;
+	volatile long          ref;
 };
